Tweaked yaml file format and added file validation.
Added validation via pydantic for the known_other_fields.yaml file. The
easiest way I could find to parse this was to have the file return a list
of dictionaries (as opposed to a set of nested dictionaries), so the format
of the file has been tweaked ever so slightly.