Now using pkg_resources for known_other_fields.
Created a pkg_data directory in yt that will contain all of the
known_other_fields.yaml files for each frontend. Added this directory
as package_data in setup.py so that these files are properly included
with yt. Modified enzo's fields.py in order to use the pkg_resources
library to load in the enzo_known_other_fields.yaml file.