Added local-dir to init files.
When running on CI, the ytcfg variable test_data_dir retains its
default value of "/does/not/exist". This causes the attempted creation
of that directory to fail. An option, "local-dir" has been added to
each pytest ini file to specify where the answer should be (or are)
saved. "local-dir" is the same variable name used by nose. Now, in
pytest_configure, if we detect the default value for the answer_dir,
then we check for an ini file value. If found, we use that. If not,
pytest will fail.

Additionally, calls to pytest.fail in requires_ds and such have been
changed to pytest.skip. This is to facilitate the running of the
"minimal" tests on CI where the data for only certain answer tests
is stored. This is akin to how it's currently done with nose.